add a default action, default for menu_order
If the last argument to _add_path() is not a coderef, we generate an empty
anonymous sub to use in its place.

repeating menu_order for every action would get old quickly, so we specify a
default one (+10 since the last default one) so the entries appear in order on
the menu.